Remove redundant sample, rename package, add README.md to card samples
