+++ collect/src/main/java/com/vgscollect/androiddemo/samples/card/README.md
@@ -1,0 +1,7 @@
+# Card Number Field Configuration
+
+This set of samples shows how to use main card number field features.
+
+## Support
+
+If you've found an error in this sample, please submit an issue or write on email: support@verygoodsecurity.com.
