+++ collect/src/main/java/com/vgscollect/androiddemo/samples/card/adapter/README.md
