+++ collect/src/main/java/com/vgscollect/androiddemo/samples/card/custom/README.md
