+++ collect/src/main/java/com/vgscollect/androiddemo/samples/card/mask/CardMaskAdapterActivity.kt
@@ -1,20 +1,17 @@
-package com.vgscollect.androiddemo.samples.brands.valid
+package com.vgscollect.androiddemo.samples.card.mask
 
 import android.os.Bundle
 import android.view.Gravity
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.setPadding
 import com.verygoodsecurity.vgscollect.core.VGSCollect
-import com.verygoodsecurity.vgscollect.util.extension.toCardBrand
-import com.verygoodsecurity.vgscollect.view.card.BrandParams
-import com.verygoodsecurity.vgscollect.view.card.CardBrand
 import com.verygoodsecurity.vgscollect.view.card.CardType
-import com.verygoodsecurity.vgscollect.view.card.validation.payment.ChecksumAlgorithm
+import com.verygoodsecurity.vgscollect.view.card.formatter.CardMaskAdapter
 import com.verygoodsecurity.vgscollect.widget.VGSCardNumberEditText
 import com.vgscollect.androiddemo.R
 import kotlinx.android.synthetic.main.activity_layout.*
 
-class ValidBrandsActivity : AppCompatActivity() {
+class CardMaskAdapterActivity : AppCompatActivity() {
 
     private lateinit var vgsCollect: VGSCollect
 
@@ -38,21 +35,18 @@ class ValidBrandsActivity : AppCompatActivity() {
         // Subscribe view
         vgsCollect.bindView(vgsEtCardNumber)
 
-        // Create custom brand
-        val brandParams = BrandParams(
-            "## #### ## #### #### ###",
-            ChecksumAlgorithm.LUHN,
-            arrayOf(19),
-            arrayOf(3)
-        )
-        val cardBrand = CardBrand(
-            "^99",
-            "Custom Brand",
-            R.drawable.ic_visa_dark,
-            brandParams
-        )
+        // Add card mask adapter and override VISA mask
+        vgsEtCardNumber.setCardMaskAdapter(object : CardMaskAdapter() {
 
-        // Set valid card brands
-        vgsEtCardNumber.setValidCardBrands(cardBrand, CardType.VISA.toCardBrand())
+            override fun getMask(
+                cardType: CardType,
+                name: String,
+                bin: String,
+                mask: String
+            ): String = when (cardType) {
+                CardType.VISA_ELECTRON -> "###### ###### ####"
+                else -> super.getMask(cardType, name, bin, mask)
+            }
+        })
     }
 }
