--- collect/src/main/java/com/vgscollect/androiddemo/samples/paymentcardnumber/README.md
@@ -1,12 +1,0 @@
-# Payment Card Number Field Sample
-
-This sample of Activity demonstrates how to use Payment Card Number field.
-
-## Prerequisites
-
-- VGS Collect SDK v.1.1.0
-- Android SDK 19
-
-## Support
-
-If you've found an error in this sample, please submit an issue or write on email: support@verygoodsecurity.com.
